Batch queue is now a submodule of batch connection
