+++ lib/riemannx/connections/batch.ex
@@ -30,7 +30,7 @@ defmodule Riemannx.Connections.Batch do
 
   import Riemannx.Settings
   import Kernel, except: [send: 2]
-  alias Riemannx.Connections.BatchQueue
+  alias Riemannx.Connections.Batch.Queue
   alias Riemannx.Proto.Msg
   alias __MODULE__
   use GenServer
@@ -60,13 +60,13 @@ defmodule Riemannx.Connections.Batch do
 
   def init(_) do
     Process.send_after(self(), :flush, batch_interval())
-    {:ok, %Batch{queue: BatchQueue.new()}}
+    {:ok, %Batch{queue: Queue.new()}}
   end
 
   def handle_cast({:push, event}, state) do
     state = %Batch{queue: queue} = push(state, event)
 
-    if BatchQueue.batch_available?(queue),
+    if Queue.batch_available?(queue),
       do: {:noreply, flush(state)},
       else: {:noreply, state}
   end
@@ -109,11 +109,11 @@ defmodule Riemannx.Connections.Batch do
     # the queue can grow larger than the configured batch size while we're waiting;
     # if the remaining part is still big enough to flush, we'll do it right after
     # this flush proc exits
-    {:ok, {remaining_queue, batch}} = BatchQueue.get_batch(queue)
+    {:ok, {remaining_queue, batch}} = Queue.get_batch(queue)
     ref = flush(batch)
     %Batch{
       state
-      | pending_flush: BatchQueue.batch_available?(queue),
+      | pending_flush: Queue.batch_available?(queue),
         ongoing_flush: ref != nil,
         flush_ref: ref,
         queue: remaining_queue,
@@ -129,7 +129,7 @@ defmodule Riemannx.Connections.Batch do
     do: %Batch{state | ongoing_flush: false, flush_ref: nil}
 
   defp push(state = %Batch{queue: queue}, event),
-    do: %Batch{state | queue: BatchQueue.push(queue, event)}
+    do: %Batch{state | queue: Queue.push(queue, event)}
 
   defp do_spawn(batch) do
     {_, ref} =
@@ -142,4 +142,89 @@ defmodule Riemannx.Connections.Batch do
 
     ref
   end
+
+  defmodule Queue do
+    @moduledoc """
+    Queue implementation for the batch module
+
+    Internally it maintains two queues:
+
+    - buffer: Contains the latest elements being inserted before the
+    size of a batch is reached
+    - batches: It has the ordered list of batches ready to be sent as
+    required
+
+    The main idea is that batches are created as the queue is filled up.
+    In this way, it avoids problems if the amount of events to be sent is
+    too big as everything is precalculated on a smaller queue (buffer)
+    """
+
+    alias Riemannx.Settings
+    alias __MODULE__
+
+    defstruct [
+      {:buffer, Qex.new()},
+      {:buffer_size, 0},
+      {:batches, Qex.new()}
+    ]
+
+    # ===========================================================================
+    # API
+    # ===========================================================================
+    def new, do: %Queue{}
+
+    def push(%{buffer: buffer, buffer_size: size} = queue, event) do
+      nqueue = %{
+        queue |
+        buffer: Qex.push(buffer, event),
+        buffer_size: size + 1
+      }
+
+      create_batch_maybe(nqueue)
+    end
+
+    def get_batch(%{buffer: buffer, batches: batches} = queue) do
+      case Qex.pop(batches) do
+        {:empty, _} ->
+          batch = Enum.to_list(buffer)
+          nqueue = %{
+            queue |
+            buffer: Qex.new(),
+            buffer_size: 0
+          }
+          {:ok, {nqueue, batch}}
+        {{:value, batch}, nbatches} ->
+          nqueue = %{
+          queue |
+          batches: nbatches
+        }
+          {:ok, {nqueue, batch}}
+      end
+    end
+
+    def batch_available?(%{batches: batches}), do: not Enum.empty?(batches)
+
+    # ===========================================================================
+    # Private
+    # ===========================================================================
+    defp create_batch_maybe(nqueue),
+      do: create_batch_maybe(nqueue, Settings.batch_size())
+
+
+    defp create_batch_maybe(
+      %{buffer_size: size, buffer: buffer, batches: batches} = queue,
+      bsize
+    ) when size >= bsize do
+      %{
+        queue |
+        buffer: Qex.new(),
+        buffer_size: 0,
+        batches: Qex.push(batches, Enum.to_list(buffer))
+      }
+    end
+
+    defp create_batch_maybe(queue, _bsize), do: queue
+
+  end
+
 end
